// src/theme/types.ts
export type Breakpoints = string[];

export type MediaQueries = {
  xs: string;
  sm: string;
  md: string;
  lg: string;
  xl: string;
  xxl : string;
  nav: string;
  maxWidthXS: string;
  maxWidthSM: string;
  maxWidthMD: string;
  maxWidthLG: string;
  maxWidthXL: string;
  maxWidthXXL: string;

};

export type Spacing = number[];

export type Radii = {
  small: string;
  default: string;
  card: string;
  circle: string;
};

export type Shadows = {
  level1: string;
  level2: string;
  level3: string;
  level4: string;
  level5: string;
  level6: string;
};

export type Gradients = {
  // bubblegum?: string;
  // cardHeader?: string;
  // blue?: string;
  multiColor: string;
  multiColor2?: string;
  multiColor3 :string;
  blue: string;
  marketCard : string;
  verticalLine :string;
  buttonBorderDark: string;
  buttonBorderLight: string;
  buttonDark: string;
  buttonLight: string;
  whiteGrayGradient: string;
  garyWhiteGradinet: string

};

export type Colors = {
  primary: string;
  primaryBright: string;
  primaryDark: string;
  secondary: string;
  success: string;
  failure: string;
  lightFailure : string;
  warning: string;
  yellow: string;
  background: string;
  backgroundDisabled: string;
  backgroundAlt: string;
  white: string;
  gray: string;
  navbarBorderColor: string;
  lavender: string;
  lightBackground: string;
  backgroundDropdownDark: string;
  seeGreen: string;
  lightSeeGreen: string;
  purple :string; 
  modalBackground : string;
  plumb : string;

};

export type ZIndices = {
  dropdown: number;
  modal: number;
};

export type Fonts = {
  xs: number;
  sm: number;
  md: number;
  lg: number;
  xl: number;
  fontSize13: number;
  fontSize14: number;
  fontSize15: number;
  fontSize16: number;
  fontSize18: number;
  fontSize20: number;
  fontSize21: number;
  fontSize22: number;
  fontSize24: number;
  fontSize26: number;

};